Moved module functions in subdirectory.
