+++ src/pnmpi/app_startup.c
@@ -37,6 +37,7 @@
 #include <pnmpi/debug_io.h>
 #include <pnmpi/private/attributes.h>
 #include <pnmpi/private/config.h>
+#include <pnmpi/private/modules.h>
 #include <pnmpi/private/mpi_interface.h>
 #include <pnmpi/private/pmpi.h>
