+++ src/pnmpi/core.h
@@ -153,10 +153,6 @@ extern pnmpi_functions_t pnmpi_function_ptrs;
 #define NOT_ACTIVATED(id) (!IS_ACTIVATED(id))
 
 void *find_symbol(const module_def_p module, const char *symbol_name);
-int pnmpi_hook_activated(const char *hook);
-void pnmpi_call_hook(const char *hook);
-
-int pnmpi_max_module_threading_level();
 
 void pnmpi_print_banner();
